Prelim: separate out flipAssets and update input amounts logic out of Swap Controller
